Overall Changes  and improvements to the files, resolved dependecies between files much more cleanly. No multiple  dependency inheritance. No majoy updates,
We need to add form function for every single suffix.
