+++ data/sample_decomposed.txt
@@ -1,77 +1,1 @@
-bu 
-ne+locative_den+aplicative_le 
-bilim+aplicative_le 
-uğraşanlar+passive_in
-buraya 
-gel+infinitive_mak
-istemeler+nounifier_i 
-doğaldı.
-ayrı+relative_ce
-bağda+active_t
-bu
-dönem+ablative_de
-türk
-nüfuz+nounifier_i 
-altına
-girm+nounifier_iş
-ve
-halifeler+nounifier_i
-ayak+ablative_de
-tuta+passive_in
-da
-burada+marking_ki
-türklerd+nounifier_i
-
-
-bu 
-ne+locative_den+aplicative_le
-bil+nounifier_im+aplicative_le
-uğraşanla+active_ir+passive_in
-buray+adverbial_e
-ge+passive_il+infinitive_mak
-iste+infinitive_me+plural_ler+accusative
-doğaldı.
- 
-ayr+nounifier_i+relative_ce
-bağda+active_t
-bu 
-döne+nounifier_im+ablative_de
-türk 
-nüfuz+nounifier_i 
-al+active_t+passive_in+adverbial_e
-
-girm+nounifier_iş
-v+adverbial_e
-halifele+active_ir+nounifier_i
-ayak+active_t+adverbial_e
-tu+active_t+factative_en
-da 
-burad+adverbial_e+marking_ki 
-türklerd+nounifier_i
-
-
-
-
-bu 
-ne+locative_den+aplicative_le
-bil+nounifier_im+aplicative_le 
-uğraş+factative_en+plural_ler+posessive_in
-bur+adverbial_e+dative_e
-gel+infinitive_mak 
-iste+infinitive_me+plural_ler+accusative 
-doğaldı.
-
-ayrı+relative_ce 
-bağda+active_t
-bu
-dönem+ablative_de
-türk 
-nüfuz+accusative 
-al+active_t+passive_in+adverbial_e
-gir+pastfactative_miş ve halife+plural_ler+accusative ay+toolative_ek+ablative_de tuta+passive_in da bura+ablative_de+marking_ki türklerdi
-
-
-
-
-
-
+bu ne+ablative_de+posessive_in+aplicative_le bilim+aplicative_le uğraş+dative_e+posessive_in+plural_ler+posessive_in bura+dative_e gelmek isteme+aplicative_le+factative_ir+accusative doğaldı. ayrıca bağ+onomatopea_de+active_it bu dönem+onomatopea_de türk nüfuz+accusative alt+posessive_in+dative_e gir+pastfactative_miş ve halife+aplicative_le+factative_ir+accusative ayakta tuta+posessive_in da burada+marking_ki türklerdi
