+++ ml_ranking_model.py
@@ -1,605 +1,408 @@
 """
 Improved Machine Learning Ranking Model for Turkish Morphological Decomposition
 
-Major improvements:
-    - Fixed configuration bugs and parameter handling
-    - True batch processing for 5-10x speed improvement
-    - Better loss functions with triplet margin
-    - Data augmentation for negative examples
-    - Early stopping and learning rate finder
-    - Optimized architecture options
+Key improvements:
+    - Removed external dependencies - works with pre-encoded data
+    - More concise code without losing features
+    - Better type hints and documentation
+    - Optimized batch processing
+    - Streamlined training loop
 """
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from typing import List, Tuple, Optional, Dict
-import math
-
-# Assuming these exist in your project
-import util.suffixes as sfx
-from data.config import TrainingConfig
 
 # ============================================================================
-# IMPROVED NEURAL NETWORK MODEL
+# NEURAL NETWORK MODEL
 # ============================================================================
 
-class ImprovedRanker(nn.Module):
-    """
-    Enhanced Transformer-based ranker.
-    
-    Improvements:
-        - Concatenated embeddings instead of sum (more expressive)
-        - Optional LSTM alternative for speed
-        - Better initialization
-    """
+
+#TODO triplets ve lsm her zaman false olacak\ duzelt.
+class Ranker(nn.Module):
+    """Transformer or LSTM-based ranker for morphological decompositions."""
     
     def __init__(
         self, 
+        suffix_vocab_size: int,
+        num_categories: int = 2,
         embed_dim: int = 128,
         num_layers: int = 4,
         num_heads: int = 8,
-        num_categories: int = 2,
         use_lstm: bool = False,
         dropout: float = 0.1
     ):
-        """
-        Initialize the improved ranker.
-        
-        Args:
-            embed_dim: Dimensionality of embeddings
-            num_layers: Number of encoder layers
-            num_heads: Number of attention heads (Transformer only)
-            num_categories: Number of POS categories (Noun/Verb)
-            use_lstm: Use LSTM instead of Transformer (faster)
-            dropout: Dropout probability
-        """
         super().__init__()
         self.embed_dim = embed_dim
-        self.num_layers = num_layers
-        self.num_heads = num_heads
         self.use_lstm = use_lstm
         
-        # Embeddings (we'll concatenate, so total = 3 * embed_dim)
-        self.suffix_embed = nn.Embedding(
-            len(sfx.ALL_SUFFIXES) + 1, 
-            embed_dim, 
-            padding_idx=0
-        )
-        self.category_embed = nn.Embedding(
-            num_categories + 1, 
-            embed_dim, 
-            padding_idx=0
-        )
+        # Embeddings (concatenated: suffix + category + position)
+        self.suffix_embed = nn.Embedding(suffix_vocab_size + 1, embed_dim, padding_idx=0)
+        self.category_embed = nn.Embedding(num_categories + 1, embed_dim, padding_idx=0)
         self.pos_embed = nn.Embedding(50, embed_dim)
         
-        # Project concatenated embeddings to model dimension
-        self.input_projection = nn.Linear(embed_dim * 3, embed_dim)
+        # Project concatenated embeddings
+        self.input_proj = nn.Linear(embed_dim * 3, embed_dim)
         
-        # Encoder (Transformer or LSTM)
+        # Encoder
         if use_lstm:
             self.encoder = nn.LSTM(
-                embed_dim,
-                embed_dim // 2,
-                num_layers=num_layers,
+                embed_dim, embed_dim // 2, num_layers=num_layers,
                 dropout=dropout if num_layers > 1 else 0,
-                bidirectional=True,
-                batch_first=True
+                bidirectional=True, batch_first=True
             )
         else:
-            encoder_layer = nn.TransformerEncoderLayer(
-                d_model=embed_dim,
-                nhead=num_heads,
-                dim_feedforward=embed_dim * 4,
-                dropout=dropout,
-                batch_first=True,
-                activation='gelu'  # Better than ReLU
+            layer = nn.TransformerEncoderLayer(
+                d_model=embed_dim, nhead=num_heads,
+                dim_feedforward=embed_dim * 4, dropout=dropout,
+                batch_first=True, activation='gelu'
             )
-            self.encoder = nn.TransformerEncoder(encoder_layer, num_layers=num_layers)
+            self.encoder = nn.TransformerEncoder(layer, num_layers=num_layers)
         
         # Scoring head
         self.scorer = nn.Sequential(
-            nn.Linear(embed_dim, embed_dim),
-            nn.LayerNorm(embed_dim),
-            nn.GELU(),
-            nn.Dropout(dropout),
-            nn.Linear(embed_dim, embed_dim // 2),
-            nn.GELU(),
-            nn.Dropout(dropout),
+            nn.Linear(embed_dim, embed_dim), nn.LayerNorm(embed_dim), nn.GELU(), nn.Dropout(dropout),
+            nn.Linear(embed_dim, embed_dim // 2), nn.GELU(), nn.Dropout(dropout),
             nn.Linear(embed_dim // 2, 1)
         )
         
         self._init_weights()
     
-    def _init_weights(self) -> None:
-        """Initialize weights with Xavier/Kaiming."""
+    def _init_weights(self):
         for name, p in self.named_parameters():
             if p.dim() > 1:
-                if 'embed' in name:
-                    nn.init.xavier_uniform_(p)
-                else:
-                    nn.init.kaiming_normal_(p, mode='fan_out', nonlinearity='relu')
+                nn.init.xavier_uniform_(p) if 'embed' in name else nn.init.kaiming_normal_(p)
             elif 'bias' in name:
                 nn.init.zeros_(p)
     
-    def forward(
-        self, 
-        suffix_ids: torch.Tensor, 
-        category_ids: torch.Tensor,
-        mask: Optional[torch.Tensor] = None
-    ) -> torch.Tensor:
+    def forward(self, suffix_ids: torch.Tensor, category_ids: torch.Tensor, 
+                mask: Optional[torch.Tensor] = None) -> torch.Tensor:
         """
-        Forward pass.
-        
         Args:
             suffix_ids: (batch, seq_len)
             category_ids: (batch, seq_len)
-            mask: (batch, seq_len) - True = padding
-            
+            mask: (batch, seq_len) - True for padding
         Returns:
             scores: (batch,)
         """
-        batch_size, seq_len = suffix_ids.shape
+        B, L = suffix_ids.shape
         
-        # Create position indices
-        pos = torch.arange(seq_len, device=suffix_ids.device)
-        pos = pos.unsqueeze(0).expand(batch_size, seq_len)
+        # Embeddings
+        pos = torch.arange(L, device=suffix_ids.device).unsqueeze(0).expand(B, L)
+        x = torch.cat([
+            self.suffix_embed(suffix_ids),
+            self.category_embed(category_ids),
+            self.pos_embed(pos)
+        ], dim=-1)
+        x = self.input_proj(x)
         
-        # Get embeddings and concatenate
-        suffix_emb = self.suffix_embed(suffix_ids)
-        category_emb = self.category_embed(category_ids)
-        pos_emb = self.pos_embed(pos)
-        
-        x = torch.cat([suffix_emb, category_emb, pos_emb], dim=-1)
-        x = self.input_projection(x)
-        
-        # Apply encoder
+        # Encode
         if self.use_lstm:
             x, _ = self.encoder(x)
         else:
             x = self.encoder(x, src_key_padding_mask=mask)
         
-        # Mean pooling
+        # Mean pool (mask-aware)
         if mask is not None:
-            mask_expanded = mask.unsqueeze(-1).expand_as(x)
-            x = x.masked_fill(mask_expanded, 0.0)
-            seq_len_count = (~mask).sum(dim=1, keepdim=True).float()
-            pooled = x.sum(dim=1) / seq_len_count.clamp(min=1)
+            x = x.masked_fill(mask.unsqueeze(-1), 0.0)
+            pooled = x.sum(dim=1) / (~mask).sum(dim=1, keepdim=True).float().clamp(min=1)
         else:
             pooled = x.mean(dim=1)
         
-        # Score
-        score = self.scorer(pooled).squeeze(-1)
-        
-        return score
+        return self.scorer(pooled).squeeze(-1)
 
 
 # ============================================================================
-# DATA AUGMENTATION
+# DATA UTILITIES
 # ============================================================================
 
 class DataAugmenter:
     """Generate negative examples for contrastive learning."""
     
     @staticmethod
-    def generate_negatives(
-        valid_decomposition: Tuple[str, List],
-        num_negatives: int = 3
-    ) -> List[Tuple[str, List]]:
+    def corrupt_chain(suffix_chain: List[Tuple[int, int]], num_negatives: int = 3) -> List[List[Tuple[int, int]]]:
         """
-        Generate negative examples by corrupting valid decomposition.
+        Corrupt a valid suffix chain to create negative examples.
         
-        Strategies:
-            1. Remove random suffix
-            2. Swap adjacent suffixes
-            3. Replace suffix with incompatible one
+        Args:
+            suffix_chain: List of (suffix_id, category_id) tuples
+            num_negatives: Number of corrupted versions to generate
+            
+        Returns:
+            List of corrupted chains
         """
-        root, suffix_chain = valid_decomposition
+        if not suffix_chain:
+            return []
+        
         negatives = []
-        
-        if len(suffix_chain) == 0:
-            return negatives
         
         # Strategy 1: Remove random suffix
         if len(suffix_chain) > 1:
             for i in range(min(num_negatives, len(suffix_chain))):
                 corrupted = suffix_chain[:i] + suffix_chain[i+1:]
                 if corrupted:
-                    negatives.append((root, corrupted))
+                    negatives.append(corrupted)
         
         # Strategy 2: Swap adjacent suffixes
         if len(suffix_chain) > 1:
-            for i in range(len(suffix_chain) - 1):
+            for i in range(min(num_negatives - len(negatives), len(suffix_chain) - 1)):
                 corrupted = suffix_chain[:i] + [suffix_chain[i+1], suffix_chain[i]] + suffix_chain[i+2:]
-                negatives.append((root, corrupted))
+                negatives.append(corrupted)
         
-        # Strategy 3: Duplicate a suffix (unlikely to be valid)
-        if len(suffix_chain) > 0:
-            corrupted = suffix_chain + [suffix_chain[0]]
-            negatives.append((root, corrupted))
+        # Strategy 3: Duplicate a suffix
+        if len(negatives) < num_negatives and suffix_chain:
+            negatives.append(suffix_chain + [suffix_chain[0]])
         
         return negatives[:num_negatives]
 
 
 # ============================================================================
-# IMPROVED TRAINING MANAGER
+# TRAINER
 # ============================================================================
 
-class ImprovedTrainer:
-    """
-    Enhanced trainer with true batching and better optimization.
-    
-    Improvements:
-        - Real batch processing
-        - Triplet loss option
-        - Early stopping
-        - Learning rate finder
-        - Gradient clipping
-        - Better logging
-    """
+class Trainer:
+    """Enhanced trainer with batching, early stopping, and flexible loss."""
     
     def __init__(
         self, 
-        model: ImprovedRanker,
+        model: Ranker,
+        model_path: str = "data/turkish_morph_model.pt",
         lr: float = 1e-4,
         weight_decay: float = 0.01,
         batch_size: int = 32,
-        device: str = None,
+        device: Optional[str] = None,
         patience: int = 10,
         use_triplet_loss: bool = False
     ):
-        """Initialize improved trainer."""
         self.model = model
         self.batch_size = batch_size
         self.device = device or ('cuda' if torch.cuda.is_available() else 'cpu')
-        self.model = self.model.to(self.device)
+        self.model.to(self.device)
         self.patience = patience
         self.use_triplet_loss = use_triplet_loss
+        self.path = model_path
         
-        # Optimizer with proper learning rate
         self.optimizer = torch.optim.AdamW(
-            model.parameters(), 
-            lr=lr,
-            weight_decay=weight_decay,
-            betas=(0.9, 0.999)
+            model.parameters(), lr=lr, weight_decay=weight_decay, betas=(0.9, 0.999)
         )
-        
-        # Cosine annealing scheduler (better than ReduceLROnPlateau)
         self.scheduler = torch.optim.lr_scheduler.CosineAnnealingWarmRestarts(
-            self.optimizer,
-            T_0=10,  # Restart every 10 epochs
-            T_mult=2,
-            eta_min=lr * 0.01
+            self.optimizer, T_0=10, T_mult=2, eta_min=lr * 0.01
         )
         
-        # Training tracking
-        self.training_history: List[float] = []
-        self.validation_history: List[float] = []
+        # Training state
+        self.train_history: List[float] = []
+        self.val_history: List[float] = []
         self.best_val_loss = float('inf')
-        self.epochs_without_improvement = 0
+        self.epochs_no_improve = 0
         self.global_step = 0
+        
+        # Try loading checkpoint
         try:
-            self.load_checkpoint(TrainingConfig.model_path)
-            print(f"✅ Loaded existing model from {TrainingConfig.model_path}")
-        except Exception as e:
-            print(f"⚠️  Could not load model: {e}")
+            self.load_checkpoint(self.path)
+            print(f"✅ Loaded model from {self.path}")
+        except:
+            print(f"⚠️  Starting fresh (no checkpoint found)")
     
-    def prepare_batch(
-        self,
-        examples: List[Tuple[str, List[List], int]]
-    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
+    def _prepare_batch(self, examples: List[Tuple[List[Tuple[int, int]], List[List[Tuple[int, int]]], int]]) \
+            -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, List[int]]:
         """
-        Prepare a true batch of multiple training examples.
+        Prepare batch from examples.
         
         Args:
-            examples: List of (root, suffix_chains, correct_idx)
-            
+            examples: List of (root_chain, candidate_chains, correct_idx)
+                     where chains are lists of (suffix_id, category_id) tuples
+                     
         Returns:
-            Tuple of batched tensors ready for model
+            (suffix_ids, category_ids, masks, labels, examples_per_word)
         """
-        all_suffix_ids = []
-        all_category_ids = []
-        all_labels = []
+        all_suffix_ids, all_category_ids, all_labels = [], [], []
+        examples_per_word = []
         
-        for root, suffix_chains, correct_idx in examples:
-            for idx, chain in enumerate(suffix_chains):
-                # Encode suffix chain
-                obj_ids, cat_ids = sfx.encode_suffix_chain(chain)
+        for root_chain, candidates, correct_idx in examples:
+            examples_per_word.append(len(candidates))
+            
+            for idx, chain in enumerate(candidates):
+                # Unzip chain into separate lists
+                if chain:
+                    suf_ids, cat_ids = zip(*chain)
+                else:
+                    suf_ids, cat_ids = [], []
                 
-                all_suffix_ids.append(torch.tensor(obj_ids))
-                all_category_ids.append(torch.tensor(cat_ids))
+                all_suffix_ids.append(torch.tensor(suf_ids, dtype=torch.long))
+                all_category_ids.append(torch.tensor(cat_ids, dtype=torch.long))
                 all_labels.append(1.0 if idx == correct_idx else 0.0)
         
-        # Find max length and pad
-        max_len = max(len(ids) for ids in all_suffix_ids)
+        # Pad to max length
+        max_len = max(len(s) for s in all_suffix_ids) if all_suffix_ids else 1
         
-        padded_suffix_ids = []
-        padded_category_ids = []
-        masks = []
+        padded_suf, padded_cat, masks = [], [], []
+        for suf, cat in zip(all_suffix_ids, all_category_ids):
+            pad_len = max_len - len(suf)
+            padded_suf.append(F.pad(suf, (0, pad_len), value=0))
+            padded_cat.append(F.pad(cat, (0, pad_len), value=0))
+            masks.append(torch.cat([torch.zeros(len(suf), dtype=torch.bool), 
+                                   torch.ones(pad_len, dtype=torch.bool)]))
         
-        for suffix_ids, category_ids in zip(all_suffix_ids, all_category_ids):
-            pad_len = max_len - len(suffix_ids)
-            
-            padded_suffix = F.pad(suffix_ids, (0, pad_len), value=0)
-            padded_category = F.pad(category_ids, (0, pad_len), value=0)
-            
-            mask = torch.cat([
-                torch.zeros(len(suffix_ids), dtype=torch.bool),
-                torch.ones(pad_len, dtype=torch.bool)
-            ])
-            
-            padded_suffix_ids.append(padded_suffix)
-            padded_category_ids.append(padded_category)
-            masks.append(mask)
-        
-        # Stack into tensors
-        suffix_ids_batch = torch.stack(padded_suffix_ids).to(self.device)
-        category_ids_batch = torch.stack(padded_category_ids).to(self.device)
-        masks_batch = torch.stack(masks).to(self.device)
-        labels_batch = torch.tensor(all_labels, dtype=torch.float32).to(self.device)
-        
-        return suffix_ids_batch, category_ids_batch, masks_batch, labels_batch
+        return (
+            torch.stack(padded_suf).to(self.device),
+            torch.stack(padded_cat).to(self.device),
+            torch.stack(masks).to(self.device),
+            torch.tensor(all_labels, dtype=torch.float32).to(self.device),
+            examples_per_word
+        )
     
-    def compute_loss(
-        self,
-        scores: torch.Tensor,
-        labels: torch.Tensor,
-        examples_per_word: List[int]
-    ) -> torch.Tensor:
-        """
-        Compute contrastive or triplet loss.
+    def _compute_loss(self, scores: torch.Tensor, labels: torch.Tensor, 
+                     examples_per_word: List[int]) -> torch.Tensor:
+        """Compute contrastive or triplet loss."""
+        losses = []
+        start = 0
         
-        Args:
-            scores: (batch,) model scores
-            labels: (batch,) 1.0 for correct, 0.0 for incorrect
-            examples_per_word: Number of candidates per word
-        """
-        if self.use_triplet_loss:
-            # Triplet loss: anchor=correct, positive=correct, negative=worst_incorrect
-            losses = []
-            start_idx = 0
+        for count in examples_per_word:
+            end = start + count
+            word_scores = scores[start:end]
+            word_labels = labels[start:end]
             
-            for num_examples in examples_per_word:
-                end_idx = start_idx + num_examples
-                word_scores = scores[start_idx:end_idx]
-                word_labels = labels[start_idx:end_idx]
+            if self.use_triplet_loss:
+                pos_idx = (word_labels == 1.0).nonzero(as_tuple=True)[0]
+                neg_idx = (word_labels == 0.0).nonzero(as_tuple=True)[0]
                 
-                # Find positive and negative
-                positive_idx = torch.where(word_labels == 1.0)[0]
-                negative_indices = torch.where(word_labels == 0.0)[0]
-                
-                if len(positive_idx) > 0 and len(negative_indices) > 0:
-                    positive_score = word_scores[positive_idx[0]]
-                    negative_score = word_scores[negative_indices].min()
-                    
-                    # Margin ranking loss
-                    loss = F.relu(negative_score - positive_score + 1.0)
-                    losses.append(loss)
-                
-                start_idx = end_idx
-            
-            return torch.stack(losses).mean() if losses else torch.tensor(0.0, device=self.device)
-        
-        else:
-            # Contrastive softmax loss
-            losses = []
-            start_idx = 0
-            
-            for num_examples in examples_per_word:
-                end_idx = start_idx + num_examples
-                word_scores = scores[start_idx:end_idx]
-                word_labels = labels[start_idx:end_idx]
-                
-                # Softmax over candidates
-                probs = F.softmax(word_scores / 0.7, dim=0)  # Temperature 0.7
-                
-                # Cross-entropy with correct label
+                if len(pos_idx) > 0 and len(neg_idx) > 0:
+                    pos_score = word_scores[pos_idx[0]]
+                    neg_score = word_scores[neg_idx].min()
+                    losses.append(F.relu(neg_score - pos_score + 1.0))
+            else:
+                # Softmax cross-entropy
+                probs = F.softmax(word_scores / 0.7, dim=0)
                 correct_prob = (probs * word_labels).sum()
-                loss = -torch.log(correct_prob + 1e-8)
-                losses.append(loss)
-                
-                start_idx = end_idx
+                losses.append(-torch.log(correct_prob + 1e-8))
             
-            return torch.stack(losses).mean()
+            start = end
+        
+        return torch.stack(losses).mean() if losses else torch.tensor(0.0, device=self.device)
     
-    def train_epoch(
-        self,
-        training_data: List[Tuple[str, List[List], int]]
-    ) -> float:
-        """
-        Train for one epoch with true batching.
-        
-        Args:
-            training_data: List of (root, suffix_chains, correct_idx)
-            
-        Returns:
-            Average loss for the epoch
-        """
+    def train_epoch(self, data: List) -> float:
+        """Train for one epoch."""
         self.model.train()
-        epoch_losses = []
+        losses = []
         
-        # Create batches
-        num_batches = math.ceil(len(training_data) / self.batch_size)
-        
-        for i in range(num_batches):
-            start_idx = i * self.batch_size
-            end_idx = min(start_idx + self.batch_size, len(training_data))
-            batch_examples = training_data[start_idx:end_idx]
+        for i in range(0, len(data), self.batch_size):
+            batch = data[i:i + self.batch_size]
+            suf_ids, cat_ids, masks, labels, counts = self._prepare_batch(batch)
             
-            # Prepare batch
-            suffix_ids, category_ids, masks, labels = self.prepare_batch(batch_examples)
+            scores = self.model(suf_ids, cat_ids, masks)
+            loss = self._compute_loss(scores, labels, counts)
             
-            # Track examples per word for loss computation
-            examples_per_word = [len(suffix_chains) for _, suffix_chains, _ in batch_examples]
-            
-            # Forward pass
-            scores = self.model(suffix_ids, category_ids, masks)
-            
-            # Compute loss
-            loss = self.compute_loss(scores, labels, examples_per_word)
-            
-            # Backward pass
             self.optimizer.zero_grad()
             loss.backward()
-            
-            # Gradient clipping
             torch.nn.utils.clip_grad_norm_(self.model.parameters(), max_norm=1.0)
-            
-            # Update weights
             self.optimizer.step()
             
-            epoch_losses.append(loss.item())
+            losses.append(loss.item())
             self.global_step += 1
         
-        avg_loss = sum(epoch_losses) / len(epoch_losses) if epoch_losses else 0.0
-        self.training_history.append(avg_loss)
-        
+        avg_loss = sum(losses) / len(losses) if losses else 0.0
+        self.train_history.append(avg_loss)
         return avg_loss
     
-    def validate(
-        self,
-        validation_data: List[Tuple[str, List[List], int]]
-    ) -> Tuple[float, float]:
-        """
-        Validate model on held-out data.
-        
-        Returns:
-            (average_loss, accuracy)
-        """
+    def validate(self, data: List) -> Tuple[float, float]:
+        """Validate and return (loss, accuracy)."""
         self.model.eval()
-        val_losses = []
-        correct_predictions = 0
-        total_predictions = 0
+        losses, correct, total = [], 0, 0
         
         with torch.no_grad():
-            for root, suffix_chains, correct_idx in validation_data:
-                # Prepare single example
-                suffix_ids, category_ids, masks, _ = self.prepare_batch(
-                    [(root, suffix_chains, correct_idx)]
-                )
+            for example in data:
+                suf_ids, cat_ids, masks, labels, counts = self._prepare_batch([example])
+                scores = self.model(suf_ids, cat_ids, masks)
                 
-                # Get scores
-                scores = self.model(suffix_ids, category_ids, masks)
+                loss = self._compute_loss(scores, labels, counts)
+                losses.append(loss.item())
                 
-                # Compute loss
-                labels = torch.zeros(len(suffix_chains), device=self.device)
-                labels[correct_idx] = 1.0
-                
-                loss = self.compute_loss(scores, labels, [len(suffix_chains)])
-                val_losses.append(loss.item())
-                
-                # Check accuracy
-                predicted_idx = scores.argmax().item()
-                if predicted_idx == correct_idx:
-                    correct_predictions += 1
-                total_predictions += 1
+                pred_idx = scores.argmax().item()
+                if labels[pred_idx] == 1.0:
+                    correct += 1
+                total += 1
         
-        avg_loss = sum(val_losses) / len(val_losses) if val_losses else 0.0
-        accuracy = correct_predictions / total_predictions if total_predictions > 0 else 0.0
-        
-        self.validation_history.append(avg_loss)
+        avg_loss = sum(losses) / len(losses) if losses else 0.0
+        accuracy = correct / total if total > 0 else 0.0
+        self.val_history.append(avg_loss)
         
         return avg_loss, accuracy
     
-    def train(
-        self,
-        training_data: List[Tuple[str, List[List], int]],
-        validation_data: Optional[List[Tuple[str, List[List], int]]] = None,
-        num_epochs: int = 100,
-        verbose: bool = True
-    ) -> Dict:
-        """
-        Full training loop with early stopping.
-        
-        Returns:
-            Dictionary with training history
-        """
+    def train(self, train_data: List, val_data: Optional[List] = None, 
+              num_epochs: int = 100, verbose: bool = True) -> Dict:
+        """Full training loop with early stopping."""
         for epoch in range(num_epochs):
-            # Train
-            train_loss = self.train_epoch(training_data)
+            train_loss = self.train_epoch(train_data)
             
-            # Validate
-            if validation_data:
-                val_loss, val_acc = self.validate(validation_data)
+            if val_data:
+                val_loss, val_acc = self.validate(val_data)
                 
-                # Early stopping check
                 if val_loss < self.best_val_loss:
                     self.best_val_loss = val_loss
-                    self.epochs_without_improvement = 0
+                    self.epochs_no_improve = 0
+                    self.save_checkpoint(self.path)
                 else:
-                    self.epochs_without_improvement += 1
+                    self.epochs_no_improve += 1
                 
                 if verbose:
-                    print(f"Epoch {epoch+1}/{num_epochs} - "
-                          f"Train Loss: {train_loss:.4f} - "
-                          f"Val Loss: {val_loss:.4f} - "
-                          f"Val Acc: {val_acc:.4f}")
+                    print(f"Epoch {epoch+1}/{num_epochs} | Train: {train_loss:.4f} | "
+                          f"Val: {val_loss:.4f} | Acc: {val_acc:.4f}")
                 
-                # Early stopping
-                if self.epochs_without_improvement >= self.patience:
+                if self.epochs_no_improve >= self.patience:
                     print(f"Early stopping at epoch {epoch+1}")
                     break
             else:
                 if verbose:
-                    print(f"Epoch {epoch+1}/{num_epochs} - Train Loss: {train_loss:.4f}")
+                    print(f"Epoch {epoch+1}/{num_epochs} | Train: {train_loss:.4f}")
             
-            # Update learning rate
             self.scheduler.step()
         
         return {
-            'training_history': self.training_history,
-            'validation_history': self.validation_history,
+            'train_history': self.train_history,
+            'val_history': self.val_history,
             'best_val_loss': self.best_val_loss
         }
     
-    def predict(
-        self,
-        root: str,
-        suffix_chains: List[List]
-    ) -> Tuple[int, List[float]]:
+    def predict(self, candidates: List[List[Tuple[int, int]]]) -> Tuple[int, List[float]]:
         """
-        Predict best decomposition.
+        Predict best candidate.
         
+        Args:
+            candidates: List of suffix chains (each chain is list of (suffix_id, category_id))
+            
         Returns:
             (best_idx, all_scores)
         """
         self.model.eval()
         
         with torch.no_grad():
-            suffix_ids, category_ids, masks, _ = self.prepare_batch(
-                [(root, suffix_chains, 0)]
-            )
+            # Create dummy example
+            suf_ids, cat_ids, masks, _, _ = self._prepare_batch([([], candidates, 0)])
+            scores = self.model(suf_ids, cat_ids, masks)
             
-            scores = self.model(suffix_ids, category_ids, masks)
-            scores_list = scores.cpu().tolist()
-            best_idx = scores.argmax().item()
-        
-        return best_idx, scores_list
+        return scores.argmax().item(), scores.cpu().tolist()
     
-    def save_checkpoint(self, path: str) -> None:
-        """Save complete training state."""
+    def save_checkpoint(self, path: str):
         torch.save({
-            'model_state_dict': self.model.state_dict(),
-            'optimizer_state_dict': self.optimizer.state_dict(),
-            'scheduler_state_dict': self.scheduler.state_dict(),
-            'training_history': self.training_history,
-            'validation_history': self.validation_history,
+            'model_state': self.model.state_dict(),
+            'optimizer_state': self.optimizer.state_dict(),
+            'scheduler_state': self.scheduler.state_dict(),
+            'train_history': self.train_history,
+            'val_history': self.val_history,
             'best_val_loss': self.best_val_loss,
             'global_step': self.global_step
         }, path)
-        print(f"✓ Model checkpoint saved to {path}")
+        print(f"✓ Saved to {path}")
     
-    def load_checkpoint(self, path: str) -> None:
-        """Load complete training state."""
-        checkpoint = torch.load(path, map_location=self.device)
-        
-        self.model.load_state_dict(checkpoint['model_state_dict'])
-        self.optimizer.load_state_dict(checkpoint['optimizer_state_dict'])
-        self.scheduler.load_state_dict(checkpoint['scheduler_state_dict'])
-        self.training_history = checkpoint.get('training_history', [])
-        self.validation_history = checkpoint.get('validation_history', [])
-        self.best_val_loss = checkpoint.get('best_val_loss', float('inf'))
-        self.global_step = checkpoint.get('global_step', 0)
-        
-        print(f"✓ Model checkpoint loaded from {path}")
-        print(f"  Resumed at step {self.global_step}")
+    def load_checkpoint(self, path: str):
+        ckpt = torch.load(path, map_location=self.device)
+        self.model.load_state_dict(ckpt['model_state'])
+        self.optimizer.load_state_dict(ckpt['optimizer_state'])
+        self.scheduler.load_state_dict(ckpt['scheduler_state'])
+        self.train_history = ckpt.get('train_history', [])
+        self.val_history = ckpt.get('val_history', [])
+        self.best_val_loss = ckpt.get('best_val_loss', float('inf'))
+        self.global_step = ckpt.get('global_step', 0)
+        print(f"✓ Loaded from {path} (step {self.global_step})")
+
