+++ util/suffixes.py
@@ -1,256 +1,234 @@
-# suffixes.py
-
 from enum import Enum
-from typing import List, Tuple, Dict
+from typing import List, Tuple
 
 import util.word_methods as wrd
 from util.rules.suffix_rules import validate_suffix_addition as validate
+#import util.rules.suffix_forms as forms 
 
-class HasMajorHarmony(Enum):      
-    Yes = 0                      
-    No = 1                       
+class Type(Enum):
+    NOUN = 'noun'
+    VERB = 'verb'
+
+class HasMajorHarmony(Enum):
+    Yes = 0
+    No = 1
+
 
 class HasMinorHarmony(Enum):
     Yes = 0
     No = 1
 
-# Registry to auto-collect all suffixes
+
 _SUFFIX_REGISTRY = []
 
-class Suffix():
-    def __init__(self, name, suffix, comes_to, makes, major_harmony=None, minor_harmony=None, needs_y_buffer=False):
+
+class Suffix:
+    def __init__(self, name, suffix, comes_to, makes, form_function=None, major_harmony=None, minor_harmony=None, needs_y_buffer=False):
+        self.name = name
         self.suffix = str(suffix)
-        self.comes_to       = comes_to
-        self.makes          = makes
-        self.name           = name
-        self.needs_y_buffer = needs_y_buffer  # New attribute for y-buffer requirement
-        self.major_harmony  = major_harmony
-        self.minor_harmony  = minor_harmony
-        
-        # Auto-register this suffix
+        self.comes_to = comes_to
+        self.makes = makes
+        self.major_harmony = major_harmony
+        self.minor_harmony = minor_harmony
+        self.needs_y_buffer = needs_y_buffer
+        self.form_function = form_function if form_function else self._default_form
         _SUFFIX_REGISTRY.append(self)
     
     def form(self, word):
-        """Generate harmonized suffix form for given word"""
-        result_list = []
-        result = self.suffix
-
-        if self.major_harmony == HasMajorHarmony.Yes:
-            word_harmony = wrd.major_harmony(word)
-            if word_harmony == wrd.MajorHarmony.BACK:
-                # Replace front vowels with back vowels
-                result = result.replace("e", "a")
-                result = result.replace("i", "ı")
-                result = result.replace("ü", "u")
-                result = result.replace("ö", "o")
-
-        # Minor vowel harmony (round/unround)
-        if self.minor_harmony == HasMinorHarmony.Yes:
-            word_harmony = wrd.minor_harmony(word)
-            
-            if word_harmony == wrd.MinorHarmony.BACK_ROUND:
-                # o, u -> change ı to u
-                result = result.replace("ı", "u")
-            elif word_harmony == wrd.MinorHarmony.FRONT_ROUND:
-                # ö, ü -> change i to ü
-                result = result.replace("i", "ü")
-
-
-
-                # Consonant hardening: g->k, d->t, c->ç when word ends in hard consonant
-        if word[-1] in wrd.HARD_CONSONANTS and result and result[0] in ['g','c','d']:
-            if result[0] == 'g':
-                result = 'k' + result[1:]
-            elif result[0] == 'd':
-                result = 't' + result[1:]
-            elif result[0] == 'c':
-                result = 'ç' + result[1:]
-        # Y-buffer: Add 'y' before vowel-initial suffixes when word ends in vowel
-        # This applies to specific suffixes that need y-buffer (marked with #)
-            
-        result_list.append(result)
-        if word and  word[-1] in wrd.VOWELS and result[0] in wrd.VOWELS and result:
-            if self.needs_y_buffer:
-                result_list.append('y'+result)
-                result_list.append('ğ'+result)
+        return self.form_function(word, self)
+    
+    @staticmethod
+    def _default_form(word, suffix_obj):
+        result = suffix_obj.suffix
+        
+        result = Suffix._apply_major_harmony(word, result, suffix_obj.major_harmony)
+        result = Suffix._apply_minor_harmony(word, result, suffix_obj.minor_harmony)
+        result = Suffix._apply_consonant_hardening(word, result)
+        
+        result_list = [result]
+        
+        if Suffix._should_add_buffer_variants(word, result):
+            if suffix_obj.needs_y_buffer:
+                result_list.append('y' + result)
+                result_list.append('ğ' + result)
             if len(result) > 1:
                 result_list.append(result[1:])
- 
+        
         return result_list
+    
+    @staticmethod
+    def _apply_major_harmony(word, result, major_harmony):
+        if major_harmony != HasMajorHarmony.Yes:
+            return result
+        
+        if wrd.major_harmony(word) == wrd.MajorHarmony.BACK:
+            result = result.replace("e", "a")
+            result = result.replace("i", "ı")
+            result = result.replace("ü", "u")
+            result = result.replace("ö", "o")
+        
+        return result
+    
+    @staticmethod
+    def _apply_minor_harmony(word, result, minor_harmony):
+        if minor_harmony != HasMinorHarmony.Yes:
+            return result
+        
+        word_harmony = wrd.minor_harmony(word)
+        
+        if word_harmony == wrd.MinorHarmony.BACK_ROUND:
+            result = result.replace("ı", "u")
+        elif word_harmony == wrd.MinorHarmony.FRONT_ROUND:
+            result = result.replace("i", "ü")
+        
+        return result
+    
+    @staticmethod
+    def _apply_consonant_hardening(word, result):
+        if not word or not result:
+            return result
+        
+        if word[-1] not in wrd.HARD_CONSONANTS:
+            return result
+        
+        first_char = result[0]
+        if first_char not in ['g', 'c', 'd', 'ğ']:
+            return result
+        
+        hardening_map = {'g': 'k', 'd': 't', 'c': 'ç', 'ğ': 'k'}
+        return hardening_map.get(first_char, first_char) + result[1:]
+    
+    @staticmethod
+    def _should_add_buffer_variants(word, result):
+        return (word and result and 
+                word[-1] in wrd.VOWELS and 
+                result[0] in wrd.VOWELS)
+
+# v2v
+reflexive_is = Suffix("reflexive_is", "iş", Type.VERB, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+reflexive_ik = Suffix("reflexive_ik", "ik", Type.VERB, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+active_it = Suffix("active_it", "it", Type.VERB, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+active_t = Suffix("active_t", "t", Type.VERB, Type.VERB, major_harmony=HasMajorHarmony.No, minor_harmony=HasMinorHarmony.No)
+active_tir = Suffix("active_tir", "dir", Type.VERB, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+active_ir = Suffix("active_ir", "ir", Type.VERB, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+passive_il = Suffix("passive_il", "il", Type.VERB, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+passive_in = Suffix("passive_in", "in", Type.VERB, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+negative_me = Suffix("negative_me", "me", Type.VERB, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+
+# n2n
+dative_archaic_ke = Suffix("dative_archaic_ke", "ke", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+locative_den = Suffix("locative_den", "den", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+dative_e = Suffix("dative_e", "e", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No, needs_y_buffer=True)
+ablative_de = Suffix("ablative_de", "de", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+plural_ler = Suffix("plural_ler", "ler", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+counting_er = Suffix("counting_er", "er", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+cooperative_daş = Suffix("cooperative_daş", "daş", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+relative_ce = Suffix("relative_ce", "ca", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+relative_sel = Suffix("relative_sel", "sel", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+dimunitive_ek_archaic = Suffix("dimunitive_ek_archaic", "ek", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+approximative_imtrak = Suffix("approximative_imtrak", "trak", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.No, minor_harmony=HasMinorHarmony.No)
+accusative = Suffix("accusative", "i", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes, needs_y_buffer=True)
+pluralizer_archaic_iz = Suffix("pluralizer_archaic_iz", "iz", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+posessive_im = Suffix("posessive_im", "im", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+posessive_in = Suffix("posessive_in", "in", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+dimunitive_cik = Suffix("dimunitive_cik", "cik", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+actor_ci = Suffix("actor_ci", "ci", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+ordinal_inci = Suffix("ordinal_inci", "inci", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+approximative_si = Suffix("approximative_si", "si", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+privative_siz = Suffix("privative_siz", "siz", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+philicative_cil = Suffix("philicative_cil", "cil", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+composessive_li = Suffix("composessive_li", "li", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+suitative_lik = Suffix("suitative_lik", "lik", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+adjectiative_ay = Suffix("adjectiative_ay", "ay", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+marking_ki = Suffix("marking_ki", "ki", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.No, minor_harmony=HasMinorHarmony.No)
+temporative_leyin = Suffix("temporative_leyin", "leyin", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+ideologicative_izm = Suffix("ideologicative_izm", "izm", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.No, minor_harmony=HasMinorHarmony.No)
+locative_le = Suffix("locative_le", "le", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+eventative_tay = Suffix("eventative_tay", "tay", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.No, minor_harmony=HasMinorHarmony.No)
+scientist_olog = Suffix("scientist_olog", "olog", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.No, minor_harmony=HasMinorHarmony.No)
+confactuous_le = Suffix("confactuous_le", "le", Type.NOUN, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+
+# n2v
+absentative_se = Suffix("absentative_se", "se", Type.NOUN, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+onomatopea_de = Suffix("onomatopea_de", "de", Type.NOUN, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+verbifier_e = Suffix("verbifier_e", "e", Type.NOUN, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+aplicative_le = Suffix("aplicative_le", "le", Type.NOUN, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+grantative_let = Suffix("grantative_let", "let", Type.NOUN, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+transformative_les = Suffix("transformative_les", "leş", Type.NOUN, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+acquirative_len = Suffix("acquirative_len", "len", Type.NOUN, Type.VERB, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
 
 
-# ============================================================================
-# SUFFIX DEFINITIONS
-# ============================================================================
-## this part needs heavy upgrades/ the form function is not enough. We have to insert versions for most suffixes
-# V2V suffixes (Verb to Verb)
-reflexive_is            = Suffix("reflexive_is", "iş", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-reflexive_ik            = Suffix("reflexive_ik", "ik", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-active_it               = Suffix("active_it", "it", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.Yes) 
-active_t                = Suffix("active_t", "t", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.No, HasMinorHarmony.No)
-active_tir              = Suffix("active_tir", "dir", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-active_ir               = Suffix("active_ir", "ir", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-passive_il              = Suffix("passive_il", "il", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-passive_in              = Suffix("passive_in", "in", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-negative_me             = Suffix("negative_me", "me", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.No)
-
-# N2N suffixes (Noun to Noun)
-dative_archaic_ke       = Suffix("dative_archaic_ke", "ke", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-locative_den            = Suffix("locative_den", "den", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-dative_e                = Suffix("dative_e", "e", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No, needs_y_buffer=True)  # Added y-buffer
-ablative_de             = Suffix("ablative_de", "de", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-plural_ler              = Suffix("plural_ler", "ler", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-counting_er             = Suffix("counting_er", "er", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-cooperative_daş         = Suffix("cooperative_daş", "daş", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-relative_ce             = Suffix("relative_ce", "ca", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-relative_sel            = Suffix("relative_sel", "sel", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-dimunitive_ek_archaic   = Suffix("dimunitive_ek_archaic", "ek", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-approximative_imtrak    = Suffix("approximative_imtrak", "trak", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.No, HasMinorHarmony.No)
-accusative              = Suffix("accusative", "i", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes, needs_y_buffer=True)  # Added y-buffer
-pluralizer_archaic_iz   = Suffix("pluralizer_archaic_iz", "iz", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-posessive_im            = Suffix("posessive_im", "im", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-posessive_in            = Suffix("posessive_in", "in", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-dimunitive_cik          = Suffix("dimunitive_cik", "cik", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-actor_ci                = Suffix("actor_ci", "ci", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-ordinal_inci            = Suffix("ordinal_inci", "inci", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-approximative_si        = Suffix("approximative_si", "si", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-privative_siz           = Suffix("privative_siz", "siz", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-philicative_cil         = Suffix("philicative_cil", "cil", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-composessive_li         = Suffix("composessive_li", "li", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-suitative_lik           = Suffix("suitative_lik", "lik", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-adjectiative_ay         = Suffix("adjectiative_ay", "ay", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-marking_ki              = Suffix("marking_ki", "ki", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.No, HasMinorHarmony.No)
-temporative_leyin       = Suffix("temporative_leyin", "leyin", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes) ## buraya bak
-ideologicative_izm      = Suffix("ideologicative_izm", "izm", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.No, HasMinorHarmony.No) ## buraya bak
-locative_le             = Suffix("locative_le", "le", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No) ## buraya bak
-eventative_tay          = Suffix("eventative_tay", "tay", wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.No, HasMinorHarmony.No) ## buraya bak
-scientist_olog          = Suffix("scientist_olog","olog",wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.No, HasMinorHarmony.No)
-confactuous_le          = Suffix("confactuous_le","le",wrd.Type.NOUN, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No) ## yle bicimini hallet
-# N2V suffixes (Noun to Verb)
-absentative_se          = Suffix("absentative_se", "se", wrd.Type.NOUN, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.No)
-onomatopea_se           = Suffix("onomatopea_de", "de", wrd.Type.NOUN, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.No)
-verbifier_e             = Suffix("verbifier_e", "e", wrd.Type.NOUN, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.No)
-aplicative_le           = Suffix("aplicative_le", "le", wrd.Type.NOUN, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.No)
-grantative_let          = Suffix("grantative_let", "let", wrd.Type.NOUN, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.No)
-transformative_les      = Suffix("transformative_les", "leş", wrd.Type.NOUN, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.No)
-acquirative_len         = Suffix("acquirative_len", "len", wrd.Type.NOUN, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.No)
-reflexive_ik            = Suffix("reflexive_ik", "ik", wrd.Type.VERB, wrd.Type.VERB, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-
-# V2N suffixes (Verb to Noun)
-nounifier_ecek          = Suffix("nounifier_ecek" , "ecek", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No, needs_y_buffer=True)  # Added y-buffer
-infinitive_me           = Suffix("infinitive_me"  , "me", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-infinitive_mak          = Suffix("infinitive_mak" , "mek", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-factative_en            = Suffix("factative_en"   , "en", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No, needs_y_buffer=True)  # Added y-buffer
-factative_er            = Suffix("factative_er", "er", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-factative_ir            = Suffix("factative_ir", "ir", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-toolative_ek            = Suffix("toolative_ek", "ek", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-adverbial_e             = Suffix("adverbial_e", "e", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No, needs_y_buffer=True)  # Added y-buffer
-constofactative_gen     = Suffix("constofactative_gen", "gen", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-constofactative_gin     = Suffix("constofactative_gin", "gin", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-nounifier_iş            = Suffix("nounifier_iş", "iş", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes, needs_y_buffer=True)  # Added y-buffer
-perfectative_ik         = Suffix("perfectative_ik", "ik", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes, needs_y_buffer=True)  # Added y-buffer
-nounifier_i             = Suffix("nounifier_i", "i", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes, needs_y_buffer=True)
-nounifier_gi            = Suffix("nounifier_gi", "gi", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-nounifier_ge            = Suffix("nounifier_ge", "ge", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-nounifier_im            = Suffix("nounifier_im", "im", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-nounifier_in            = Suffix("nounifier_in", "in", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-nounifier_inç           = Suffix("nounifier_inç", "inç", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-nounifier_inti          = Suffix("nounifier_inti", "inti", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-toolifier_geç           = Suffix("toolifier_geç", "geç", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-toolifier_eç            = Suffix("toolifier_eç", "eç", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-neverfactative_mez      = Suffix("neverfactative_mez", "mez", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-pastfactative_miş       = Suffix("pastfactative_miş", "miş", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.Yes)
-adjectiative_ay         = Suffix("adjectiative_ay", "ay", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-
-nounifier_anak          = Suffix("nounifier_anak", "anak", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-nounifier_amak          = Suffix("nounifier_amak", "amak", wrd.Type.VERB, wrd.Type.NOUN, HasMajorHarmony.Yes, HasMinorHarmony.No)
-# V2N  inflection   
+# v2n
+nounifier_ecek = Suffix("nounifier_ecek", "ecek", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No, needs_y_buffer=True)
+infinitive_me = Suffix("infinitive_me", "me", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+infinitive_mak = Suffix("infinitive_mak", "mek", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+factative_en = Suffix("factative_en", "en", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No, needs_y_buffer=True)
+factative_er = Suffix("factative_er", "er", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+factative_ir = Suffix("factative_ir", "ir", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+toolative_ek = Suffix("toolative_ek", "ek", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+adverbial_e = Suffix("adverbial_e", "e", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No, needs_y_buffer=True)
+constofactative_gen = Suffix("constofactative_gen", "gen", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+constofactative_gin = Suffix("constofactative_gin", "gin", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+nounifier_iş = Suffix("nounifier_iş", "iş", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes, needs_y_buffer=True)
+perfectative_ik = Suffix("perfectative_ik", "ik", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes, needs_y_buffer=True)
+nounifier_i = Suffix("nounifier_i", "i", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes, needs_y_buffer=True)
+nounifier_gi = Suffix("nounifier_gi", "gi", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+nounifier_ge = Suffix("nounifier_ge", "ge", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+nounifier_im = Suffix("nounifier_im", "im", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+nounifier_in = Suffix("nounifier_in", "in", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+nounifier_inç = Suffix("nounifier_inç", "inç", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+nounifier_inti = Suffix("nounifier_inti", "inti", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+toolifier_geç = Suffix("toolifier_geç", "geç", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+toolifier_eç = Suffix("toolifier_eç", "eç", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+neverfactative_mez = Suffix("neverfactative_mez", "mez", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+pastfactative_miş = Suffix("pastfactative_miş", "miş", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.Yes)
+adjectiative_ay_v2n = Suffix("adjectiative_ay", "ay", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+nounifier_anak = Suffix("nounifier_anak", "anak", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
+nounifier_amak = Suffix("nounifier_amak", "amak", Type.VERB, Type.NOUN, major_harmony=HasMajorHarmony.Yes, minor_harmony=HasMinorHarmony.No)
 
 
-# ============================================================================
-# AUTO-COLLECTED SUFFIX LISTS
-# ============================================================================
-
-# All suffixes in order of definition
 ALL_SUFFIXES = _SUFFIX_REGISTRY
 
-# Organized by transition type for the mapping
 SUFFIX_TRANSITIONS = {
     'noun': {
-        'noun': [s for s in ALL_SUFFIXES if s.comes_to == wrd.Type.NOUN and s.makes == wrd.Type.NOUN],
-        'verb': [s for s in ALL_SUFFIXES if s.comes_to == wrd.Type.NOUN and s.makes == wrd.Type.VERB]
+        'noun': [s for s in ALL_SUFFIXES if s.comes_to == Type.NOUN and s.makes == Type.NOUN],
+        'verb': [s for s in ALL_SUFFIXES if s.comes_to == Type.NOUN and s.makes == Type.VERB]
     },
     'verb': {
-        'noun': [s for s in ALL_SUFFIXES if s.comes_to == wrd.Type.VERB and s.makes == wrd.Type.NOUN],
-        'verb': [s for s in ALL_SUFFIXES if s.comes_to == wrd.Type.VERB and s.makes == wrd.Type.VERB]
+        'noun': [s for s in ALL_SUFFIXES if s.comes_to == Type.VERB and s.makes == Type.NOUN],
+        'verb': [s for s in ALL_SUFFIXES if s.comes_to == Type.VERB and s.makes == Type.VERB]
     }
 }
 
-suffix_to_id = {}
-id_to_suffix = {}
-category_to_id: Dict[str, int] = {
-            'Noun': 0,
-            'Verb': 1
-        }
-for idx, suffix in enumerate(ALL_SUFFIXES):
-    suffix_to_id[suffix.name] = idx
-    id_to_suffix[idx] = suffix.name
+suffix_to_id = {suffix.name: idx for idx, suffix in enumerate(ALL_SUFFIXES)}
+id_to_suffix = {idx: suffix.name for idx, suffix in enumerate(ALL_SUFFIXES)}
+category_to_id = {'Noun': 0, 'Verb': 1}
 
-    
-def encode_suffix_chain( suffix_objects: List) -> Tuple[List, List]:
-    """
-    Convert a list of Suffix objects to numeric tensor representations.
-    
-    Args:
-        suffix_objects: List of Suffix objects from a decomposition
-        
-    Returns:
-        Tuple of:
-            - object_ids: Tensor of suffix IDs
-            - category_ids: Tensor of POS category IDs (Noun/Verb)
-            
-    Note:
-        Empty chains return single padding tokens (ID 0)
-    """
-    # Handle empty suffix chains
+
+def encode_suffix_chain(suffix_objects: List) -> Tuple[List, List]:
     if not suffix_objects:
         return [0], [0]
     
-    object_ids = []
-    category_ids = []
-    
-    # Convert each suffix to its numeric representation
-    for suffix_obj in suffix_objects:
-        # Get suffix ID (default to 0 for unknown suffixes)
-        suffix_id = suffix_to_id.get(suffix_obj.name, 0)
-        
-        # Get category ID (Noun=0, Verb=1)
-        category_id = category_to_id.get(suffix_obj.makes.name, 0)
-        
-        object_ids.append(suffix_id)
-        category_ids.append(category_id)
+    object_ids = [suffix_to_id.get(s.name, 0) for s in suffix_objects]
+    category_ids = [category_to_id.get(s.makes.name, 0) for s in suffix_objects]
     
     return object_ids, category_ids
 
 
-
-def find_suffix_chain(word, start_pos, root, visited=None):
-    """Recursively find valid suffix chains after a root."""
+def find_suffix_chain(word, start_pos, root, visited=None): 
     if visited is None:
         visited = set()
-
+    
     state_key = (len(root), start_pos)
     if state_key in visited:
         return []
-
+    
     visited = visited | {state_key}
     rest = word[len(root):]
-
-    # Base case: fully matched word
+    
     if not rest:
         return [([], start_pos)]
-
-    # No suffix transitions available
+    
     if start_pos not in SUFFIX_TRANSITIONS:
         return []
     
@@ -264,22 +242,42 @@ def find_suffix_chain(word, start_pos, root, visited=None):
                     subchains = find_suffix_chain(word, target_pos, next_root, visited) if remaining else [([], target_pos)]
                     
                     for chain, final_pos in subchains:
-                        # RULE VALIDATION: Check if adding suffix_obj before the existing chain is valid
-                        # The chain represents suffixes that come AFTER suffix_obj
-                        # So we validate: can suffix_obj be followed by chain[0] (if chain exists)?
-                        
-                        # First, validate adding suffix_obj to an empty chain (start of suffix sequence)
                         if not validate([], suffix_obj):
                             continue
                         
-                        # If there's a following suffix, validate the transition
                         if chain and not validate([suffix_obj], chain[0]):
                             continue
                         
-                        # Both validations passed
                         results.append(([suffix_obj] + chain, final_pos))
-                    break   
+                    break
+    
     return results
-    
 
+
+##  THE ISSUE WITH THIS  CODE is that it is too specific for the way given data is set.
+#  right now we can not differentiate between at noun and at verb because they are not seperated lets say.
+#  therefore the code bends itself to adhere to those limits/ 
+# Thinking about createing nouns and verbs seperately but that would take a lotta time.
+def decompose( word: str) -> List[Tuple]:
+    """Find all possible legal decompositions of a word."""
+    if not word:
+        return []
     
+    analyses = []
+    for i in range(1, len(word) + 1):
+        root = word[:i]
+        exists_status = wrd.exists(root)
+        
+        if exists_status == 0:
+            continue
+        
+        pos = "verb" if wrd.can_be_verb(root) else "noun"
+
+        chains = (find_suffix_chain(word, "verb", root) +
+                find_suffix_chain(word, "noun", root)) if pos == "verb" \
+                else find_suffix_chain(word, "noun", root)
+
+        for chain, final_pos in chains:
+            analyses.append((root, pos, chain, final_pos))
+    
+    return analyses
